Confirmation dialog box for delete buttons. Issue #33
A delete confirmation check has been added on the 'Admin Tasks' pages; edit_categories.php, edit_items.php, manage_users.php.

The built-in javascript confirm() function is called on <form> 'onsubmit' Event attribute.

+++ edit_items.php
@@ -53,9 +53,9 @@
                     <input type="hidden" name="item_id" value="<?php echo $row["id"] ?>">
                     </form>
                     <td>
-                        <form action="edit_items.php" method="post">
+                        <form action="edit_items.php" method="post" onsubmit="return confirm('delete this item?');">
                             <input type="hidden" name="delete_item" value="<?php echo $row["name"] ?>">
-                            <input type="submit" value="delete" class="button">
+                            <input type="submit" value="delete" class="button" >
                         </form>
                     </td>
                 </tr>
